tests: verify mysql inspectdb for float field
