+++ tests/models.py
@@ -93,6 +93,7 @@ class Product(Model):
     )
     pic = fields.CharField(max_length=200)
     body = fields.TextField()
+    price = fields.FloatField()
     created_at = fields.DatetimeField(auto_now_add=True)
     is_deleted = fields.BooleanField(default=False)
