+++ tests/test_migrate.py
@@ -981,6 +981,7 @@ def test_migrate(mocker: MockerFixture):
             "ALTER TABLE `product` DROP COLUMN `uuid`",
             "ALTER TABLE `product` DROP INDEX `uuid`",
             "ALTER TABLE `product` RENAME COLUMN `image` TO `pic`",
+            "ALTER TABLE `product` ADD `price` DOUBLE NOT NULL",
             "ALTER TABLE `email` RENAME COLUMN `id` TO `email_id`",
             "ALTER TABLE `product` ADD INDEX `idx_product_name_869427` (`name`, `type_db_alias`)",
             "ALTER TABLE `email` ADD INDEX `idx_email_email_4a1a33` (`email`)",
@@ -1027,6 +1028,7 @@ def test_migrate(mocker: MockerFixture):
             "ALTER TABLE `product` ADD `uuid` INT NOT NULL UNIQUE",
             "ALTER TABLE `product` ADD UNIQUE INDEX `uuid` (`uuid`)",
             "ALTER TABLE `product` DROP INDEX `idx_product_name_869427`",
+            "ALTER TABLE `product` DROP COLUMN `price`",
             "ALTER TABLE `email` DROP INDEX `idx_email_email_4a1a33`",
             "ALTER TABLE `product` DROP INDEX `uid_product_name_869427`",
             "ALTER TABLE `product` ALTER COLUMN `view_num` DROP DEFAULT",
@@ -1074,6 +1076,7 @@ def test_migrate(mocker: MockerFixture):
             'ALTER TABLE "product" RENAME COLUMN "image" TO "pic"',
             'ALTER TABLE "product" RENAME COLUMN "is_review" TO "is_reviewed"',
             'ALTER TABLE "product" RENAME COLUMN "is_delete" TO "is_deleted"',
+            'ALTER TABLE "product" ADD "price" DOUBLE PRECISION NOT NULL',
             'ALTER TABLE "user" ALTER COLUMN "password" TYPE VARCHAR(100) USING "password"::VARCHAR(100)',
             'ALTER TABLE "user" DROP COLUMN "avatar"',
             'ALTER TABLE "user" ALTER COLUMN "longitude" TYPE DECIMAL(10,8) USING "longitude"::DECIMAL(10,8)',
@@ -1118,6 +1121,7 @@ def test_migrate(mocker: MockerFixture):
             'ALTER TABLE "product" RENAME COLUMN "pic" TO "image"',
             'ALTER TABLE "product" RENAME COLUMN "is_deleted" TO "is_delete"',
             'ALTER TABLE "product" RENAME COLUMN "is_reviewed" TO "is_review"',
+            'ALTER TABLE "product" DROP COLUMN "price"',
             'ALTER TABLE "user" ADD "avatar" VARCHAR(200) NOT NULL DEFAULT \'\'',
             'ALTER TABLE "user" ALTER COLUMN "password" TYPE VARCHAR(200) USING "password"::VARCHAR(200)',
             'ALTER TABLE "user" ALTER COLUMN "longitude" TYPE DECIMAL(12,9) USING "longitude"::DECIMAL(12,9)',
